requirements.txt
pyqt
pyqtgraph
pillow
